+++ Pais.cs
@@ -2,6 +2,6 @@ public class Pais
 {
     public int Id { get; set; }
     public string Nome { get; set; } = string.Empty;
-    public int AreaTotal { get; set; }
-    public int Populacao { get; set; }
+    public double AreaTotal { get; set; }
+    public double Populacao { get; set; }
 }
